Some code corrections.
