+++ src/components/SharedComponents/Hero.jsx
@@ -1,26 +1,6 @@
-"use client";
-
-import { Button } from "../ui/button";
-import { FaGoogle } from "react-icons/fa";
-import { useAuthContext } from "@/Contexts/AuthContexts";
-import { toast } from "sonner";
-import { useRouter } from "next/navigation";
+import UserLoginButton from "../User/UserLoginButton";
 
 function Hero() {
-  const { googleLogIn } = useAuthContext();
-
-  const router = useRouter();
-
-  const handleLogIn = async () => {
-    try {
-      await googleLogIn();
-      router.push("/profile");
-    } catch (error) {
-      console.log(error);
-      toast.warning(error.message);
-    }
-  };
-
   return (
     <section className="flex flex-col py-4 px-4 gap-4">
       <div className="text-center px-2 py-2">
@@ -32,13 +12,7 @@ function Hero() {
         </p>
       </div>
       <div className="flex w-full justify-center items-center">
-        <Button
-          className="inline-flex h-full items-center gap-4"
-          onClick={handleLogIn}
-        >
-          <FaGoogle size={15} />
-          <span className="text-lg font-mono">Login</span>
-        </Button>
+        <UserLoginButton />
       </div>
     </section>
   );
