+++ src/components/User/UserAvatar.jsx
@@ -1,10 +1,10 @@
 import React from "react";
 import { Avatar, AvatarImage } from "@/components/ui/avatar";
 
-function UserAvatar({ size, user }) {
+function UserAvatar({ size, photoUrl }) {
   return (
     <Avatar className={`${size ? size : "w-10 h-10"}`}>
-      <AvatarImage src={user?.photoURL} />
+      <AvatarImage src={photoUrl} />
     </Avatar>
   );
 }
