+++ src/components/User/UserProfileWindow.jsx
@@ -1,14 +1,17 @@
 "use client";
 import UserAvatar from "./UserAvatar";
 import UserHeader from "./UserHeader";
+import { useAuthContext } from "@/Contexts/AuthContexts";
 
-function UserProfileCard({ user }) {
+function UserProfileWindow() {
+  const { user } = useAuthContext();
+
   return (
     <section className="flex flex-col gap-2 rounded-sm border border-input shadow-xl h-[500px] w-[600px]">
       <UserHeader />
       <div className="flex-1 flex flex-col gap-2  items-center justify-center px-2 py-2 w-full h-full">
         <div className="flex w-full justify-center">
-          <UserAvatar user={user} size={"w-40 h-40"} />
+          <UserAvatar photoUrl={user?.photoURL} size={"w-40 h-40"} />
         </div>
         <div className="flex flex-col gap-2 text-center">
           <p className="font-mono tracking-wide text-4xl">
@@ -21,4 +24,4 @@ function UserProfileCard({ user }) {
   );
 }
 
-export default UserProfileCard;
+export default UserProfileWindow;
